Test execution report mechanism
This commit adds tests for the new execution report feature to the
existing test suite; these new tests simply make hard-coded assertions
about the number of instructions of each opcode (c.f. syscalls) that
each example program executed.

I've deliberately held back on adding these new tests _everywhere_ they
could be applied for the sake of minimising noise in the test suite /
the diff in this PR. The tests as added cover the new feature with
sufficient generality, but I'd be very happy to add the full set if the
extra coverage is desirable.