Update quickstart.md
On the first step in Option 1, I got an error which is easily fixed, but a quick update to the docs would be easy. Or perhaps from a product perspective there should be a default/optimal setting so you don't need to pass the argument. 

 % cargo prove new fibonacci

error: the following required arguments were not provided:
  <--bare|--evm>

Usage: cargo prove new <--bare|--evm> <NAME>

I propose to add the argument --evm and update the description with this line: 

This command uses the `evm` template which includes Solidity smart contracts for onchain integration.