fix: Write diagnostics to stderr (vs stdout)
Before this change (even when logging was not enabled) SP1 was writing diagnostic messages to stdout, breaking the implicit contract that only the application logic should decide what gets written to stdout.
This contract is important, e.g. for passing the output for further processing in shell-driven use cases:

```sh
cat ./input-data.json | my-app | another-app
```

This change aims to fix this issue, and to safeguard against it (to a reasonable extent) in the future.

I've left alone tests, examples and build scripts, and tried to make the change as narrow as possible and to affect only the production code paths.